fixed possible bugs/vulnerabilities from implicit conversions, explicitly defined variables that are always positive as unsigned

+++ src/headers.h
@@ -12,23 +12,23 @@ struct http_header {
     union {
         struct phr_header phr_header;
         struct {
-            const char *name;
-            size_t      name_len;
-            const char *value;
-            size_t      value_len;
+            const char  *name;
+            unsigned int name_len;
+            const char  *value;
+            unsigned int value_len;
         };
     };
 };
 
 struct header_value {
-    char *value;
-    int   value_len;
+    char        *value;
+    unsigned int value_len;
 };
 
 enum http_header_props {
-    MAX_NUM_HEADERS          = 100,
-    HEADER_VALUE_VALID       = 2,
-    HEADER_EXISTS            = 4,
+    MAX_NUM_HEADERS = 100,
+    HEADER_VALUE_VALID = 2,
+    HEADER_EXISTS = 4,
     HEADER_VALUE_EXCEEDS_MAX = 8,
 };
 
@@ -38,7 +38,7 @@ struct header_hashset;
 static inline void http_header_init(struct http_header *header,
                                     const char         *header_name,
                                     const char         *header_value) {
-    header->name  = header_name;
+    header->name = header_name;
     header->value = header_value;
 }
 
@@ -46,8 +46,9 @@ struct header_hashset *init_hashset(void);
 void                   reset_header_hashset(struct header_hashset *set);
 void                   destroy_hashset(struct header_hashset *set);
 struct header_value   *http_get_header(struct header_hashset *set,
-                                       const char *name, int name_len);
-int http_set_header(struct header_hashset *set, const char *name, int name_len,
-                    const char *value, int value_len);
+                                       const char *name, unsigned int name_len);
+int http_set_header(struct header_hashset *set, const char *name,
+                    unsigned int name_len, const char *value,
+                    unsigned int value_len);
 
 #endif /* __HEADERS_H */
